S15.d.2: extract up.ts preflight phase (1660 → 1296 LOC)
Second sub-slice of the up.ts multi-PR sub-train. Extract the
entire preflight phase (auto-detect dev mode, cached-context
prefill, banner + tool checks, Azure auth + subscription,
interactive prompts, RBAC + provider preflight, SKU availability
check, --dry-run plan print) to cli/src/commands/up/preflight.ts.

Caller in up.ts is a 4-line dispatch:
  const { runPreflight } = await import("./up/preflight.js");
  const preflightResult = await runPreflight(options);
  if (preflightResult === null) return; // --dry-run
  const { rg } = preflightResult;

Helper isValidAzureHost moved to the new module and re-exported
for the single deploy-section consumer.

Verification:
- up.ts: 1660 → 1296 LOC (further reduction in d.3, d.4)
- new up/preflight.ts: 392 LOC
- tsc --noEmit / lint (27 warnings, baseline) / build: clean
- npm test: 454 passed; 2 skipped pre-existing

Subsequent sub-slices d.3 (Helm + AgentMesh deploy) and d.4
(sandbox bring-up + summary) will continue toward the §15 800 cap.

No behavioral change: preflight body moved verbatim.

Audit: docs/security-audits/2026-04-29-phase2-hotspot-up-cli-d2.md
